+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 * esm filter: convert top level requires into imports
 * lit filter: updates from polymer LitElement to Lit v2.
-  added: customElements, es2022 syntax for properties
+  added: customElements, es2022 syntax for properties,
+  query, queryAll, queryAsync
 
 # 4.1.7 / 2021-09-20